tablas
agregue las tablas que me faltaban en la bd y arregle a update que me tiraba error

+++ database/migrations/2026_05_28_220002_add_imagen_to_productos_table.php
@@ -12,8 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('productos', function (Blueprint $table) {
-        $table->string('imagen')->nullable()->after('stock');
-    });
+            if (!Schema::hasColumn('productos', 'imagen')) {
+                $table->string('imagen')->nullable()->after('stock');
+            }
+        });
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-       Schema::table('productos', function (Blueprint $table) {
-        $table->dropColumn('imagen');
-    });
+        Schema::table('productos', function (Blueprint $table) {
+            if (Schema::hasColumn('productos', 'imagen')) {
+                $table->dropColumn('imagen');
+            }
+        });
     }
 };
